cli handling clean updated

+++ ablation_cli.py
@@ -1,4 +1,5 @@
 import shutil
+import os
 from sys import argv
 from typing import Callable
 from ablation_experiments import SELECTABLE_EXPERIMENTS
@@ -50,7 +51,7 @@ if __name__ == "__main__":
             break
 
     if (flag := "--clean") in args:
-        if custom_dir_path:
+        if custom_dir_path and os.path.exists(custom_dir_path):
             shutil.rmtree(custom_dir_path)
 
     experiment(custom_dir_path,custom_seed)
